requirements.txt
pdfplumber
openai>=1.0.0
python-dotenv
tqdm
